@@ -4,17 +4,16 @@ plugins {
     id 'com.google.gms.google-services'
     id 'kotlin-kapt'
     id "androidx.navigation.safeargs.kotlin"
-    id 'dagger.hilt.android.plugin'
-    id 'kotlin-android-extensions'
+    id 'com.google.dagger.hilt.android'
 }
 
 android {
-    compileSdk 31
+    compileSdk 33
 
     defaultConfig {
         applicationId "uz.ibroxim.uzendictionary"
         minSdk 21
-        targetSdk 31
+        targetSdk 33
         versionCode 1
         versionName '1.0.1'
 
