@@ -6,9 +6,8 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.AsyncListDiffer
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
-import kotlinx.android.synthetic.main.adapter_favorite.view.*
-import kotlinx.android.synthetic.main.adapter_word.view.*
 import uz.techie.uzendictionary.R
+import uz.techie.uzendictionary.databinding.AdapterFavoriteBinding
 import uz.techie.uzendictionary.models.Favorite
 import uz.techie.uzendictionary.models.Word
 import uz.techie.uzendictionary.utils.Constants
@@ -30,27 +29,27 @@ class FavoriteAdapter(private val listener: WordListener):RecyclerView.Adapter<F
     val differ = AsyncListDiffer(this, diffCallBack)
 
 
-    inner class WordViewHolder(itemView:View):RecyclerView.ViewHolder(itemView){}
+    inner class WordViewHolder(val binding:AdapterFavoriteBinding):RecyclerView.ViewHolder(binding.root){}
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): WordViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.adapter_favorite, parent, false)
-        return WordViewHolder(view)
+        val binding = AdapterFavoriteBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        return WordViewHolder(binding)
     }
 
     override fun onBindViewHolder(holder: WordViewHolder, position: Int) {
         val favorite = differ.currentList[position]
 
         if (primaryLang == Constants.LANG_UZBEK){
-            holder.itemView.adapter_favorite_title.text = favorite.word_uz
+            holder.binding.adapterFavoriteTitle.text = favorite.word_uz
         }
         else{
-            holder.itemView.adapter_favorite_title.text = favorite.word_en
+            holder.binding.adapterFavoriteTitle.text = favorite.word_en
         }
 
         holder.itemView.setOnClickListener {
             listener.onItemClick(favorite)
         }
-        holder.itemView.adapter_favorite_delete.setOnClickListener {
+        holder.binding.adapterFavoriteDelete.setOnClickListener {
             listener.onDeleteClick(favorite)
         }
     }
